[logging] logging all browser errors and logs to server
we need to know what is happening at the client end. Every error and every $log.info is
logged in json format. Integrated with winston logging. All logging happens in json format.
At the browser, $exceptionhandler and $log delegates logging from console to the server.

As a failsafe, we also have window.onerror

Signed-off-by: himangshu <himangshu@sokratik.com>